requirements.txt
elevenlabs==0.2.21
Flask==2.3.2
google_api_python_client==2.95.0
google_auth_oauthlib==1.0.0
protobuf==3.20.3
pydub==0.25.1
